+++ Fuse.sln
@@ -28,7 +28,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Storage", "Storage\Storage.
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Parser", "Parser\Parser.csproj", "{7AE71A6E-AABE-40D4-968B-0E6018EE376D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ImportTool", "ImportTool\ImportTool.csproj", "{15E86B0C-B2E8-4655-907E-8905B01B492E}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Import", "Import\Import.csproj", "{15E86B0C-B2E8-4655-907E-8905B01B492E}"
 	ProjectSection(ProjectDependencies) = postProject
 		{7AE71A6E-AABE-40D4-968B-0E6018EE376D} = {7AE71A6E-AABE-40D4-968B-0E6018EE376D}
 	EndProjectSection
