+++ Import/Import.csproj
@@ -7,8 +7,8 @@
     <ProjectGuid>{15E86B0C-B2E8-4655-907E-8905B01B492E}</ProjectGuid>
     <OutputType>Exe</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>ImportTool</RootNamespace>
-    <AssemblyName>ImportTool</AssemblyName>
+    <RootNamespace>Import</RootNamespace>
+    <AssemblyName>Import</AssemblyName>
     <TargetFrameworkVersion>v4.5.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
@@ -34,7 +34,7 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <PropertyGroup>
-    <StartupObject>ImportTool.Program</StartupObject>
+    <StartupObject>Import.Program</StartupObject>
   </PropertyGroup>
   <ItemGroup>
     <Reference Include="CommandLine, Version=1.9.71.2, Culture=neutral, PublicKeyToken=de6f01bd326f8c32, processorArchitecture=MSIL">
